@@ -7,7 +7,7 @@ authors = [
 description = "Predict gender using first name using data from World Gender Name Dictionary 2.0."
 readme = "README.md"
 license = { file="LICENSE" }
-requires-python = ">=3.7"
+requires-python = ">=3.9"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
